requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pydantic 
psycopg2-binary
spacy
python-dotenv
python-multipart
pgvector
openai
jinja2